The components were moved to separate files, the map method was added to the data arrays
