+++ src/components/Dialogs/Dialogs.jsx
@@ -1,50 +1,39 @@
 import React from 'react';
 import s from './Dialogs.module.css';
-import {NavLink} from 'react-router-dom';
+import DialogItem from './DialogItem/DialogItem';
+import Message from './Message/Message'
 
-const DialogItem = (props) => {
-    let path = '/dialogs/' + props.id; 
 
-    return(
-        <div className={s.dialog}><NavLink to={path}>{props.name}</NavLink></div>
-    )
-}
-
-const Message = (props) => {
-    return(
-        <div className={s.message}>{props.message}</div>
-    )
-}
 
 const Dialogs = (props) => {
 
-    let dialogsData = [
+    let dialogs = [
         {id: 1, name: 'Mike'},
         {id: 2, name: 'Anny'},
         {id: 3, name: 'Lisa'},
         {id: 4, name: 'John'},
     ];
 
-    let messagesData = [
+    let messages = [
         {id: 1, message: 'Hi'},
         {id: 2, message: 'Hi?'},
         {id: 3, message: 'Who said hi?'},
         {id: 4, message: 'Im said hi.'},
     ];
 
+    let dialogsElements = dialogs.map( d => <DialogItem name={d.name} id={d.id}/> );
+
+    let messagesElements = messages.map( m => <Message message={m.message} id={m.id}/>)
+
+    
+
     return(
         <div className={s.dialogsWrapper}>
             <div className={s.dialogs}>
-                <DialogItem name={dialogsData[0].name} id={dialogsData[0].id}/>
-                <DialogItem name={dialogsData[1].name} id={dialogsData[1].id}/>
-                <DialogItem name={dialogsData[2].name} id={dialogsData[2].id}/>
-                <DialogItem name={dialogsData[3].name} id={dialogsData[3].id}/>
+                {dialogsElements}
             </div>
             <div className={s.messages}>
-                <Message message={messagesData[0].message} id={messagesData[0].id}/>
-                <Message message={messagesData[1].message} id={messagesData[1].id}/>
-                <Message message={messagesData[2].message} id={messagesData[2].id}/>
-                <Message message={messagesData[3].message} id={messagesData[3].id}/>
+                {messagesElements}
             </div>
         </div>
     )
