+++ src/components/Dialogs/Message/Message.jsx
@@ -1,0 +1,12 @@
+import React from 'react';
+import s from './../Dialogs.module.css';
+
+
+const Message = (props) => {
+    return(
+        <div className={s.message}>{props.message}</div>
+    )
+}
+
+
+export default Message;
